project_#3_Multi-Agent Environmental Research System/requirements.txt
# =============================================================================
# Multi-Agent Environmental Research System - Dependencies
# =============================================================================

# Core LLM Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20

# Web Search
tavily-python>=0.3.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Web Framework
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
httpx>=0.26.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# Utilities
tenacity>=8.2.3
tiktoken>=0.5.2
rich>=13.7.0

# Logging
loguru>=0.7.2

# Export
python-docx>=1.1.0
markdown>=3.5.0
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Development
black>=23.12.0
isort>=5.13.0
